+++ Sources/ScrollKit/Extensions/View+RoundedScollContent.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2025 Daniel Saidi. All rights reserved.
 //
 
-#if os(iOS) || os(visionOS)
 import SwiftUI
 
 public extension View {
@@ -27,7 +26,7 @@ public extension View {
     }
 }
 
-@available(iOS 16.0, *)
+@available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *)
 public extension View {
     
     /// Make a scroll view's content view overlap the scroll
@@ -49,9 +48,6 @@ public extension View {
             .frame(maxHeight: .infinity)
             .scrollViewContentWithHeaderOverlap(overlap)
     }
-}
-
-public extension View {
         
     /// Make a scroll view header view apply rounded corners
     /// that cut out a mask for the scroll view content view.
@@ -62,34 +58,29 @@ public extension View {
     func scrollViewHeaderWithRoundedContentMask(
         _ points: Double = 0
     ) -> some View {
-        if #available(iOS 16.0, *) {
-            if points > 0 {
-                self.mask {
-                    VStack(spacing: 0) {
-                        /// Make the black color overflow waaaay up.
-                        Color.black.scaleEffect(100, anchor: .bottom)
-                        ZStack {
-                            Color.white
-                            UnevenRoundedRectangle(
-                                topLeadingRadius: 20,
-                                topTrailingRadius: 20
-                            )
-                            .fill(.black)
-                        }
-                        .compositingGroup()
-                        .luminanceToAlpha()
-                        .frame(height: 20)
+        if points > 0 {
+            self.mask {
+                VStack(spacing: 0) {
+                    /// Make the black color overflow waaaay up.
+                    Color.black.scaleEffect(100, anchor: .bottom)
+                    ZStack {
+                        Color.white
+                        UnevenRoundedRectangle(
+                            topLeadingRadius: 20,
+                            topTrailingRadius: 20
+                        )
+                        .fill(.black)
                     }
+                    .compositingGroup()
+                    .luminanceToAlpha()
+                    .frame(height: 20)
                 }
-            } else {
-                self
             }
         } else {
             self
         }
     }
 }
-#endif
 
 #Preview {